removed some stdout spew during codegen.
fixed issue w/ incorrect code generation for kokkos parallel-for where lambda constructs have
a reference type.  issue #2.
