--- kitsune-tests/kokkos/pfor_1d_sum.cpp
@@ -1,90 +1,0 @@
-/*
- * ###########################################################################
- * Copyright (c) 2019, Los Alamos National Security, LLC All rights
- * reserved. Copyright 2019. Los Alamos National Security, LLC. This
- * software was produced under U.S. Government contract DE-AC52-06NA25396
- * for Los Alamos National Laboratory (LANL), which is operated by Los
- * Alamos National Security, LLC for the U.S. Department of Energy. The
- * U.S. Government has rights to use, reproduce, and distribute this
- * software.  NEITHER THE GOVERNMENT NOR LOS ALAMOS NATIONAL SECURITY,
- * LLC MAKES ANY WARRANTY, EXPRESS OR IMPLIED, OR ASSUMES ANY LIABILITY
- * FOR THE USE OF THIS SOFTWARE.  If software is modified to produce
- * derivative works, such modified software should be clearly marked, so
- * as not to confuse it with the version available from LANL.
- *  
- * Additionally, redistribution and use in source and binary forms, with
- * or without modification, are permitted provided that the following
- * conditions are met: 1.       Redistributions of source code must
- * retain the above copyright notice, this list of conditions and the
- * following disclaimer. 2.      Redistributions in binary form must
- * reproduce the above copyright notice, this list of conditions and the
- * following disclaimer in the documentation and/or other materials
- * provided with the distribution. 3.      Neither the name of Los Alamos
- * National Security, LLC, Los Alamos National Laboratory, LANL, the U.S.
- * Government, nor the names of its contributors may be used to endorse
- * or promote products derived from this software without specific prior
- * written permission.
-  
- * THIS SOFTWARE IS PROVIDED BY LOS ALAMOS NATIONAL SECURITY, LLC AND
- * CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING,
- * BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND
- * FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL LOS
- * ALAMOS NATIONAL SECURITY, LLC OR CONTRIBUTORS BE LIABLE FOR ANY
- * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE
- * GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
- * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
- * OTHERWISE) ARISING IN ANY WAY OUT OF THE US
- * ########################################################################### 
- * 
- * Notes
- *
- * ##### 
- */
-#include <cstdio>
-#include <iostream>
-using namespace std;
-
-
-#include <Kokkos_Core.hpp>
-
-typedef Kokkos::View<double*> View;
-
-const size_t SIZE = 16;
-
-void initialize(View &v) {
-  Kokkos::parallel_for(SIZE, KOKKOS_LAMBDA(const int i) {
-    v(i) = double(i);
-  });
-}
-
-void dump(View &v) {
-  for(size_t i = 0; i < SIZE; ++i) { 
-    cout << "v(" << i << ") = " << v(i) << endl; 
-  }
-}
-
-int main (int argc, char* argv[]) {
-
-  Kokkos::initialize (argc, argv);
-
-  { // scope used for cleanup (to hush kokkos)... 
-    View a("a", SIZE);
-    View b("b", SIZE);
-    View c("c", SIZE);
-
-    initialize(a);
-    dump(a);
-    initialize(b);
-    dump(b);
-
-    Kokkos::parallel_for(SIZE, KOKKOS_LAMBDA (const int i) {
-      c(i) = a(i) + b(i);
-    });
-
-    dump(c); 
-  }
-
-  Kokkos::finalize ();
-}
